move DiskVariant
