+++ sled-agent/src/sim/storage.rs
@@ -26,7 +26,6 @@ use illumos_utils::zfs::DatasetProperties;
 use omicron_common::api::external::ByteCount;
 use omicron_common::disk::DatasetName;
 use omicron_common::disk::DiskIdentity;
-use omicron_common::disk::DiskVariant;
 use omicron_common::disk::SharedDatasetConfig;
 use omicron_uuid_kinds::DatasetUuid;
 use omicron_uuid_kinds::ExternalZpoolUuid;
@@ -37,6 +36,7 @@ use omicron_uuid_kinds::ZpoolUuid;
 use propolis_client::VolumeConstructionRequest;
 use serde::Serialize;
 use sled_agent_types::dataset::LocalStorageDatasetEnsureRequest;
+use sled_agent_types::disk::DiskVariant;
 use sled_agent_types::inventory::OmicronSledConfig;
 use sled_agent_types::inventory::ZpoolHealth;
 use sled_agent_types::support_bundle::NESTED_DATASET_NOT_FOUND;
