+++ sled-hardware/Cargo.toml
@@ -22,6 +22,7 @@ omicron-uuid-kinds.workspace = true
 rand.workspace = true
 schemars.workspace = true
 serde.workspace = true
+sled-agent-types.workspace = true
 sled-hardware-types.workspace = true
 slog.workspace = true
 slog-error-chain.workspace = true
