+++ sled-hardware/src/disk.rs
@@ -6,10 +6,11 @@ use camino::{Utf8Path, Utf8PathBuf};
 use illumos_utils::fstyp::Fstyp;
 use illumos_utils::zpool::Api;
 use illumos_utils::zpool::Zpool;
-use omicron_common::disk::{DiskIdentity, DiskVariant};
+use omicron_common::disk::DiskIdentity;
 use omicron_common::zpool_name::ZpoolName;
 use omicron_uuid_kinds::ZpoolUuid;
 use serde::{Deserialize, Serialize};
+use sled_agent_types::disk::DiskVariant;
 use slog::Logger;
 use slog::{info, warn};
 use slog_error_chain::SlogInlineError;
