+++ sled-storage/Cargo.toml
@@ -24,6 +24,7 @@ omicron-uuid-kinds.workspace = true
 schemars = { workspace = true, features = [ "chrono", "uuid1" ] }
 serde.workspace = true
 serde_json.workspace = true
+sled-agent-types.workspace = true
 sled-hardware.workspace = true
 slog.workspace = true
 slog-error-chain.workspace = true
