+++ sled-storage/zfs-test-harness/src/lib.rs
@@ -18,12 +18,12 @@ use illumos_utils::zfs::Zfs;
 use key_manager::KeyManager;
 use key_manager::StorageKeyRequester;
 use omicron_common::disk::DatasetConfig;
-use omicron_common::disk::DiskVariant;
 use omicron_common::disk::SharedDatasetConfig;
 use omicron_common::zpool_name::ZpoolName;
 use omicron_uuid_kinds::DatasetUuid;
 use omicron_uuid_kinds::ExternalZpoolUuid;
 use omicron_uuid_kinds::InternalZpoolUuid;
+use sled_agent_types::disk::DiskVariant;
 use sled_storage::config::MountConfig;
 use sled_storage::dataset::M2_DEBUG_DATASET;
 use sled_storage::disk::Disk;
